+++ src/adPEAS.ps1
@@ -148,6 +148,7 @@ if ($Script:ScriptPath) {
     . "$Script:ScriptPath\modules\Helpers\ConvertTo-SID.ps1"
     . "$Script:ScriptPath\modules\Helpers\Test-IsPrivileged.ps1"
     . "$Script:ScriptPath\modules\Helpers\Get-ObjectOwner.ps1"
+    . "$Script:ScriptPath\modules\Helpers\ConvertFrom-ADInterval.ps1"
     . "$Script:ScriptPath\modules\Helpers\ConvertFrom-GPPPassword.ps1"
     . "$Script:ScriptPath\modules\Helpers\ConvertFrom-VBE.ps1"
     . "$Script:ScriptPath\modules\Helpers\ConvertFrom-Base64OrFile.ps1"